requirements.txt
python-dotenv==1.0.0
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.18
pyotp>=2.9.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
Flask-SQLAlchemy>=3.0.0
google-api-python-client>=2.100.0
gunicorn>=21.0.0
